pyproject.toml
[tool.poetry]
name = "turtlefunt"
version = "0.1.0"
description = "Create turtlefun T-Shirts"
authors = ["Robert Schneider <robert@schneirob.de>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"
colorcet = "^3.1.0"
loguru = "^0.7.2"

[tool.poetry.scripts]
turtlefunt = "turtlefunt.turtlefunt:turtlefunt"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
coverage = {extras = ["toml"], version = "^7.6.1"}
pytest-cov = "^5.0.0"
pytest-loguru = "^0.4.0"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["turtlefunt"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
